Merged the new IO functionality written by Falko Schmidt into the trunk. For a smooth transition the new functionality is disabled per default but can be included by adding -DNEW_IO to the compiler flags (additionally the boost library is needed, which is supplied in the libs subdirectory. You may also check the BOOST_VER variable in the Makefile to ensure the version matches your boost installation).
With the new IO the Command line parameters (which are explained when calling MarDyn without any 
command line parameters) and the input file format are changed.
To convert Moldy or MDProject input files into the new format have a look into the 
tools folder, where a converter for each format exists.

git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@47 a63bd714-7e14-4b5e-94e7-302c8c8ff188