Added the working implementation of the kd-tree-based load balancing. - Changes in KDDecomposition: previous version was incomplete - Changes in ParticleContainer and derived Classes: Added methods   for getting and setting a particlePairHandler. Changed   _particlePairsHandler to be a pointer instead of a reference to be   able to easily switch between pairHandlers. - ParticlePairs2LoadCalcAdapter.h: New particlePairHandler used by the   KDDecomposition for guessing the load - changes to other classes in folder parallel: the methods exchangeMolecules   and balanceAndExchange had an unused parameter (double rc) which was unused   and therefore has been removed.
The committed code of the KDDecomposition was fully working in the branch,
but there are still some problems in the trunk. It seems to work fine
with AdaptiveSubCells, but with LinkedCells the results are not correct
(It worked fine with LinkedCells in the branch). So there is still some
bug either in LinkedCells or in KDDecomposition.

git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@204 a63bd714-7e14-4b5e-94e7-302c8c8ff188