add BlockTraverse structure to avoid duplicate traversePairs() method
This datastructure is there to prevent code duplication in LinkedCells
and AdaptiveSubCells classes.  The traversePairs() method in these
classes is very similar, except for
- naming (cells <-> subCells), etc.
- offsets (NeighbourOffsets equal for every cell <-> different for every
  cell)
This class contains a generic traversePairs() method and all needed
member variables. There are two constructors to deal with the different
datastructures for the offsets in LinkedCells and AdaptiveSubCells.

git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@232 a63bd714-7e14-4b5e-94e7-302c8c8ff188