Introduced pointer and reference to the simulation object
Introduced pointer global_simulation and reference _simulation.
The latter can be used for easy access to simulation data instead of
passing them all the time, e.g to call the global exit function.
The pointer should not be used as it will make an eventual switch to
the model view controller model much harder.
Though it is not recommended to make massive use of this access method
as it's mostly a sign for shortcomings in the implementation.

Some smaller cleanups and comments.

git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@258 a63bd714-7e14-4b5e-94e7-302c8c8ff188