Bugfix: Moved MPI and logfile initialization into main
Moved MPI initialization into main function as it makes more sense
than introducing additional parameters to the constructor of the
simulatoin class.
At the same time moved the logfile initialization and application info
output to main as we can do this after MPI_Init was called and so can
clean up the Simulation class.

git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@293 a63bd714-7e14-4b5e-94e7-302c8c8ff188