README!!!
This revision sets the foundation framework for being able to carry out "World record" simulations. A necessary change is to be able to replace the Molecules with the lightweight 1CLJ molecule implementation (which stores only position, velocity and unique ID), and still be able to compile the entire code-base without wrapping half of the code in ifdef-s.



git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@4151 a63bd714-7e14-4b5e-94e7-302c8c8ff188