README 2:
(hit "commit" by mistake, before being ready with the commitmessage)


the old class "Molecule is now called FullMolecule" and it derives from a base class "MoleculeInterface".
Please bear with us and introduce any necessary changes in "MoleculeInterface"
and provide a stub at least for compiling Molecule_WR (the compiler should force you to).

In order to avoid performance degradation, "Molecule" is now a typedef, either to "FullMolecule" or "Molecule_WR". Forward declaring a particle (writing "class Molecule;" in a header, instead of "#include molecules/Molecule.h") is now somewhat cumbersome, so I'm putting it in a separate header file "molecules/MoleculeForwardDeclaration.h".

Bottom line:
code outside of the Molecule class-es stays as is, except for forward declaration of molecules, which should go through MoleculeForwardDeclaration.h. The former code for class Molecule is now in FullMolecule.h. Plase propagate changes to the public interfaces also to MoleculeInterface and Molecule_WR.

More changes will follow, but this should suffice for the general structure.

git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@4152 a63bd714-7e14-4b5e-94e7-302c8c8ff188