introduce CavityEnsemble::countNeighbours(..).
This function is intended to replace the LinkedCells::countNeighbours, which goes through the cellProcessor-traversals. Instead, the new one uses the RegionParticleIterator, which is the interface that we want to have exposed to the outside. 

Note: in this revision it is easy to compare the old functionality LinkedCells::countNeighbours versus the new functionality CavityEnsemble::countNeighbours. Just change in LinkedCells::cavityStep(..). 

Not tested, as I don't have input for the CavityEnsemble..

git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@4510 a63bd714-7e14-4b5e-94e7-302c8c8ff188