rework ParticleIterator classes, so that they can work with the upcoming Verlet List implementation.
This is a necessary change for the moment, but might be reverted in the future, as it causes a performance loss - the ParticleIterator methods now involve virtual calls and are not inlined. To be reviewed once the Verlet Lists implementation is finished. 

Also, rename functions in Molecule.h and ParticleCell.h to avoid duplicated function names.

git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@4657 a63bd714-7e14-4b5e-94e7-302c8c8ff188