Orientation distribution function plugin added and rotating Stockmayer fluid enabled:
By adding a site of the type "Stockmayer" to a component in the xml file, the moments of inertia for the component are set to Ixx = Iyy = Izz = 1.
This input doesn't change interaction parameters but it enables the creation of a central single-site rotating fluid, which was previously not possible in ls1.
A simple LJ-center with a superimposed Dipole (i.e. Stockmayer fluid) in the same position can now rotate.
Rotation previously required at least two sites in different positions.

ODF: Calculates orientation distributions in the first coordination shell for pure dipolar fluids or binary mixtures made of dipolar fluids.
When using, choose [0 0 1] for the dipole vector in the xml file.
Mutual orientation is described by the angles Phi1, Phi2, and Gamma12.
Phi1 and Phi2 describe the angles between the between the dipole vectors of the two molecules and the connecting axis of the molecules.
Gamma12 describes the torque between the dipole vectors. (see K. Langenbach/Chemical Engineering Science 174 (2017) 40–55)
Mutual orientation frequencies are sorted into bins for the three angles in the output.
This plugin requires changes because it slows down the code drastically if the ODF is sampled frequently.