FixRegion:
Function: This plugin is supposed to keep particles within a specified volume from moving during simulation. The particles still interact with non-fixed particles via forces.
Specifically, this plugin is supposed to fix an atomistic wall, by immobilizing the bottom layers of the wall.
These bottom layers of the wall hold the rest of the wall in place via force interactions.

Implementation: Every time step the force and velocity of the particles within the specified volume is set to zero.

Problems: In test simulations of a sessile droplet on the wall, the droplet exhibits an unusual pressure profile.
I am not sure what causes this. I previously tried a similar approach as in the Dropaccelerator plugin, wherein I marked the particles of interest with a bool.
I then excluded these particles from the leapfrog integration. This approach also caused problems as the number of marked particles did not remain constant.
The current implementation however causes the aforementioned problems with the pressure.
Is this way of realizing a fixed region sensible? At which point during a time step would it make the most sense to set velocities and forces to zero (if at all)?