Add ADIOS2 option for Makefile builds
Setting the variable ADIOS2=1 for Makefile builds in the src directory
enables building with ADIOS2 support, i.e., the  AdiosWriter/Reader.
ADIOS2 headers and libraries are detected via the adios2-config tool.
If the tool is not in the PATH one can use the ADIOS2_INCDIR and
ADIOS2_LIBDIR variables to provide the necessary paths for compiling
and linking.

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>