added hack to circumvent gcc bug
When same code is tested with clang (at all optimization levels)
and with gcc optimization O0, the code runs bugfree, but with
gcc O3 a segfault occurs when the cell service is accessed. This is a
workaround pending more investigation, however it is not known whether a
better solution exists yet.