Fixed issue of identical ID numbers in VTK-output
ID numbers of molecules were identical when using VTK-writer for all output except the one of rank 0. The XML Schema has been adapted such that the basic data type (called "decimal" in the XML Schema) for storing the output data is now mapped to "long double" (was "double" before), thus allowing for longer integers to be stored.