Fix inconsistent last seek table entry in MmpldWriter output
The MMPLD file format includes a seek table for the data frames.
The last entry of this table should point at the end of the last
frame. However, this is not identical to the file size in some cases.
Correct seek table entries are already inserted by the write_frame()
method. Therefore, removing the erroneous code in the finish() method.

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>
Co-authored-by: HomesGH <homes@tu-berlin.de>