Change default initialization for Quaternions to 1 + 0i + 0j + 0k
The current initialization is not representing a pure rotation as it is not
normalized. Change this to the identity quaternion 1 + 0i + 0j + 0k, i.e.,
Q(w = 1, x = 0, y = 0, z = 0), to be consistent with the needs of the
Molecule interface.

Note: Initialisation to Q(0,0,0,0) might show some performance improvement
over Q(1,0,0,0) but tests showed neglectable influence on the overall
performance of ls1. So we stay with the safe and convenient solution.

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>