Remove mardy_assert from OptionParser to prevent initialisation order issues
MarDyn_assert uses the logger infrastructure of ls1 internally. However the
logger is initialized with command line options causing a circular
dependence. Therefore, remove mardyn_assert from the OptionParser.
If something is wrong with the command line it is early in the execution
and therefore no need for the extensive logger infrastructure to be used.

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>