Changed usage of ENABLE_PERSISTENT Macro.
The CollectiveCommunicationPersistent.h and
CollectiveCommunicationPersistent_helper.h will now always be included,
if ENABLE_MPI is set. This was done to get rid of the ifdef-endif blocks
enclosing all collective communication calls. The ENABLE_PERSISTENT was
moved inside the CollCommObj class to determine if a persistent
collective communication call is to be used.
The implementation of CollCommObj also changed because of that. Now the
persistent collective and the normal collective functions share
exactly the same interface with the only difference being the
ENABLE_PERSISTENT determining which is used. The new communicate()
functions now start all types of collective communication, therefore is
it no longer possible to omit the collective function type at
construction and give it during communication.