Refactored the CollCommObj::get function.
The function now returns a tuple of the received data. This change makes
it obvious that data is returned from this function. Another advantage
is that with structured bindings (C++ 17) the returned tuple can be used
to declare the variables into which the tuple values should be copied
(see docu in CollectiveCommunicationPersistent.h for more detail).