Refactored the MPI_Members member in CollCommObj.
Previously the MPI communicator was stored as a separate member, but it
is now part of the MPI_Members member. This was done because it makes
code clearer to store members with similar usage together. There was a
check added to the CollCommObj constructor that checks, if the given
communicator has been changed (this is especially important to pass all
tests). The MPI_Members members has been reverted to a pointer. This was
done to fix the MPI_x_free after MPI_Finalize issue.